added some more debug information

+++ 3d-raytracing-renderer/Equal.h
@@ -2,6 +2,6 @@
 #define EPSILON 0.00001
 #include <math.h>
 
-inline bool floatEqual(float a, float b) {
+inline bool floatEqual(double a, double b) {
 	return (abs(a - b) < EPSILON);
 }
